Work on tick-timing tests...
...and simplify adaptation to Keil/IAR.  Note that lptimTick.c has been
compatible with Keil and IAR for a long time now, so this commit merely
updates a comment in that file.  And including ulp.h from FreeRTOS.h is
now done in a way that is compatible with Keil and IAR (following an
established pattern).

The tick-timing tests use the RTC to analyze tick accuracy.  The code in
this commit merely monitors the timing.  There is no specific stimulus
stress test to the tick timing (yet).  Using the code in this commit, we
can now observe tick timing with the built-in tickless idle, with the
custom tickless idle in lptimTick.c, and with no tickless idle at all.
We did prove that the MSI PLL mode works like a PLL (albeit with high
jitter) as noted in #1.