Add another demo state, finish #1 and #2
Use LPTIM2 as a nuisance interrupt to stress test the tick timing. And
update ulp driver to support both STOP 1 and STOP 2 since LPTIM2
operates in STOP 1 but not in STOP 2.

Update readme.md to describe the three demo states.

Update FreeRTOSConfig.h to indicate more clearly the minimum
requirements for basic integration of lptimTick.c.