Add support for configMIN_RUN_BETWEEN_DEEP_SLEEPS
This is a workaround for erratum 2.3.21 on 'L476/'L486.  The demo
application doesn't need the workaround at the moment, as explained in
FreeRTOSConfig.h with this commit, so for now the workaround is
disabled.

This repo demonstrates integration of lptimTick.c, so it's best to
include this workaround.  The workaround is probably needed for many
real-world applications.