Handle CMPOK before CMPM in LPTIM_IRQHandler()
Consider a configuration with a tick every 10 timer counts.  When we
handle the tick at count 50, the ISR writes 60 to the CMP register.  If
the timer is at count 59 when we handle the tick at count 50, the timer
will not raise a CMPM event for the tick at count 60 because of the CMP
sync mechanism.  In this case, our design depends on the CMPOK interrupt
to trigger the tick at count 60.  If the CMPOK event occurs while we're
still in the ISR -- perhaps due to a higher priority ISR or a slow core
clock -- our design requires that we not only acknowledge the CMPOK
interrupt but that we *subsequently* check for a compare match.

Prior to this commit, our ISR checked for a compare match (a tick)
*before* acknowledging any CMPOK interrupt.  For the example above, the
ISR could clear the CMPOK interrupt (around timer count 61 or 62), and
leave the LPTIM no longer requesting an interrupt, with the tick at
count 60 now missed.  In reality, the tick at count 60 isn't actually
missed.  That's because the NVIC causes one additional execution of the
ISR in this case, and that execution identifies the tick at count 60.
The Cortex-M NVIC is sensitive to edges (not just levels) on IRQ lines.
In this example, there is an edge on the LPTIM's IRQ line in the middle
of the ISR, when the CMPOK event happens.

After this commit, the ISR checks for a CMPOK event before checking for
a compare match.  The code now more closely reflects the actual design
of the software.